# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Whatsapp Clone API Documentation",
        "contact": {
            "email": "mail@satrio.dev"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/chatrooms": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Get all chatrooms",
                "description": "Retrieves a list of all chatrooms.",
                "operationId": "ce7fe6f309c557b6e5c5b4fa2acbed05",
                "responses": {
                    "200": {
                        "description": "A list of all chatrooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "General Chat"
                                            },
                                            "max_members": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-06T12:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-06T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Create a new chatroom",
                "description": "Creates a new chatroom with a specified name and maximum number of members.",
                "operationId": "e4ba0c519a31cb0a4bfcf0031f3e7193",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "max_members"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "General Chat"
                                    },
                                    "max_members": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chatroom created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "General Chat"
                                        },
                                        "max_members": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-06T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-06T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The 'name' field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chatrooms/my": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Get all chatrooms the user belongs to",
                "description": "Fetches a list of chatrooms that the authenticated user is a member of.",
                "operationId": "29fb2dd715594fb730ceb9429cb1dc61",
                "responses": {
                    "200": {
                        "description": "List of chatrooms the user belongs to",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "General Chatroom"
                                            },
                                            "max_members": {
                                                "type": "integer",
                                                "example": 50
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-01T12:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-01T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, the user must be logged in to access this resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chatrooms/{chatroom}/enter": {
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Join a chatroom",
                "description": "Allows an authenticated user to join a chatroom if there is space and they are not already a member.",
                "operationId": "990649538e8db7d02764f28786a61ef5",
                "parameters": [
                    {
                        "name": "chatroom",
                        "in": "path",
                        "description": "ID of the chatroom to join",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "13662cf0-ead2-48b7-a056-ded77823fe9c"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully joined the chatroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Joined chatroom"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "The chatroom is full",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Chatroom is full"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "User is already a member of the chatroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User already in chatroom"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, the user must be logged in to access this resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chatrooms/{chatroom}/leave": {
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Leave a chatroom",
                "description": "Allows an authenticated user to leave a chatroom.",
                "operationId": "0fca3374da5da9644eeb25e71f7273df",
                "parameters": [
                    {
                        "name": "chatroom",
                        "in": "path",
                        "description": "ID of the chatroom to leave",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "13662cf0-ead2-48b7-a056-ded77823fe9c"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully left the chatroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Left chatroom"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, the user must be logged in to access this resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chatroom not found or user is not a member of the chatroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Chatroom not found or user is not a member"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}